+++ contracts/_refactor/impl/lib/_Configuration.sol
@@ -4,11 +4,9 @@ pragma solidity ^0.5.0;
 library _Configuration {
     struct State {
         // The percentage protocol takes from deposit interest as reserve.
-        uint protocolReserveRatio;
-
+        uint256 protocolReserveRatio;
         // Lock all functionalities related to deposit, loan and liquidating.
         bool isUserActionsLocked;
-
         address protocolAddress;
         address priceOracleAddress;
     }
@@ -16,19 +14,33 @@ library _Configuration {
     event LockUserActions();
     event UnlockUserActions();
 
-    function setPriceOracleAddress(State storage self, address priceOracleAddress) external {
-        require(priceOracleAddress != address(0), "Configuration: invalid price oracle address");
+    function setPriceOracleAddress(
+        State storage self,
+        address priceOracleAddress
+    ) external {
+        require(
+            priceOracleAddress != address(0),
+            'Configuration: invalid price oracle address'
+        );
 
         self.priceOracleAddress = priceOracleAddress;
     }
 
-    function setProtocolAddress(State storage self, address protocolAddress) external {
-        require(protocolAddress != address(0), "Configuration: invalid protocol address");
+    function setProtocolAddress(State storage self, address protocolAddress)
+        external
+    {
+        require(
+            protocolAddress != address(0),
+            'Configuration: invalid protocol address'
+        );
 
         self.protocolAddress = protocolAddress;
     }
 
-    function setProtocolReserveRatio(State storage self, uint protocolReserveRatio) external {
+    function setProtocolReserveRatio(
+        State storage self,
+        uint256 protocolReserveRatio
+    ) external {
         self.protocolReserveRatio = protocolReserveRatio;
     }
