+++ contracts/mocks/LoanManagerMock.sol
@@ -1,22 +1,19 @@
 pragma solidity ^0.5.0;
 
-import "../LoanManager.sol";
-import "../Loan.sol";
-
+import '../LoanManager.sol';
+import '../Loan.sol';
 
 contract LoanManagerMock is LoanManager {
     Loan[] public loans;
 
     function loan(
-        uint term,
-        address loanAsset, 
+        uint256 term,
+        address loanAsset,
         address collateralAsset,
-        uint loanAmount,
-        uint collateralAmount,
+        uint256 loanAmount,
+        uint256 collateralAmount,
         bool useFreedCollateral
-    ) 
-        public returns (Loan) 
-    {
+    ) public returns (Loan) {
         Loan currLoan = super.loan(
             term,
             loanAsset,
